refactor: 예상 구조로 컴포넌트 추가 및 파일 구조 변경
