+++ react/src/components/Footer.jsx
@@ -1,0 +1,4 @@
+function Footer() {
+  return;
+}
+export default Footer;
