+++ react/src/pages/additemPage.jsx
@@ -1,3 +1,4 @@
 function additemPage() {
   return null;
 }
+export default additemPage;
