+++ react/src/pages/itemsPage.module.css
@@ -1,0 +1,17 @@
+.wrapper {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+}
+.main {
+  display: flex;
+  flex-direction: column;
+  margin: 24px;
+  justify-content: center;
+  align-items: center;
+  gap: 40px;
+  max-width: 1200px;
+  min-width: 398px;
+}
